New UUID Generation library for BBDB
In the proposed file format version 7, we want to assign a unique
ID for each bbdb entry. The proposed choice is RFC4122 UUID. I have
found one reasonable and working implementation in pure emacs lisp
by Kan-Ru Chen. However, as I am not qualified to coomment on
the quality of that implementation, I figured it prudent to take
a 'hybridg' approach, namely:

- If there is a system provided uuid generator (such as uuidgen
  on MacOS X), we will use it by firing a shell comomand

- If such a utility could not be found, we will use Kan-ru's
  elisp code to get an ID

The present commit merely brings in the library into BBDB while
not referencing it anywere.